+++ Defaults.lua
@@ -3,6 +3,8 @@ local ADDON_NAME, NS = ...
 NS.ADDON_NAME = ADDON_NAME
 NS.ACCOUNT_DB_NAME = ADDON_NAME .. "DB"
 NS.CHARACTER_DB_NAME = ADDON_NAME .. "CharacterDB"
+NS.ACCOUNT_STORAGE_METADATA_KEY = "__yvBags"
+NS.ACCOUNT_STORAGE_SCHEMA_VERSION = 1
 NS.FRAME_NAME = ADDON_NAME .. "Frame"
 NS.ITEM_BUTTON_GLOBAL_NAME_PREFIX = ADDON_NAME .. "ItemListButton"
 NS.ADDON_MEDIA_PATH = "Interface\\AddOns\\" .. ADDON_NAME .. "\\Media\\"
@@ -14,6 +16,11 @@ NS.defaults = {
             replaceBlizzardBags = true,
             autosellGrayJunk = false,
         },
+        pins = {
+            items = {},
+        },
+    },
+    profile = {
         list = {
             sortKey = "quality",
             sortAscending = false,
@@ -23,7 +30,6 @@ NS.defaults = {
         },
         pins = {
             displayMode = "top",
-            items = {},
         },
         display = {
             showCooldownsInName = true,